content filtering / add :comment_filters option / rdiscount support
Filters are just helper methods that take a single text argument
and apply some transformation. The "filter" method can be used to
chain multiple filters together:

    filter text, :markdown, :smartify, :sanitize

This is equivalent to:

    sanitize(smartify(markdown(text)))

The new system is now being used for comments, entry bodies, and
summaries.

The :comment_filters option can be set to an array of filters to apply
to comments.

While here, added support for rdiscount if installed; falls back to
BlueCloth.